feat: add design system components including dynamic images, empty states, and video cards

+++ core/model/src/main/java/alireza/nezami/model/data/VideoFileDto.kt
@@ -7,14 +7,14 @@ import kotlinx.serialization.Serializable
 data class VideoFileDto(
         val url: String? = null,
         val width: Int? = null,
-        val height: Int? = null,
+        val height: Double? = null,
         val size: Int? = null,
         val thumbnail: String? = null
 ) {
     fun toDM() = VideoFileDM(
         url = url.orEmpty(),
         width = width ?: 0,
-        height = height ?: 0,
+        height = height ?: 0.0,
         size = size ?: 0,
         thumbnail = thumbnail.orEmpty()
     )
@@ -24,7 +24,7 @@ data class VideoFileDto(
         fun empty() = VideoFileDto(
             url = "",
             width = 0,
-            height = 0,
+            height = 0.0,
             size = 0,
             thumbnail = ""
         )
